+++ perf/buffer.h
@@ -27,20 +27,20 @@ extern int _trace_bufferStart(void);
 extern int _trace_bufferFinish(void);
 
 
-extern int _trace_bufferRead(u8 chan, void *buf, size_t bufsz);
+extern ssize_t _trace_bufferRead(u8 chan, void *buf, size_t bufsz);
 
 
-extern int _trace_bufferWrite(u8 chan, const void *data, size_t sz);
+extern ssize_t _trace_bufferWrite(u8 chan, const void *data, size_t sz);
 
 
 extern int _trace_bufferWaitUntilAvail(u8 chan, size_t sz);
 
 
 /* returns bytes available to write */
-extern int _trace_bufferAvail(u8 chan);
+extern ssize_t _trace_bufferAvail(u8 chan);
 
 
-extern int _trace_bufferDiscard(u8 chan, size_t sz);
+extern ssize_t _trace_bufferDiscard(u8 chan, size_t sz);
 
 
 #endif
